back/requirements.txt
numpy==1.22.2
pandas==1.4.1
fastapi==0.73.0
uvicorn==0.17.4
sqlalchemy==1.4.31
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
bcrypt==3.2.0
python-dotenv==0.19.2
pg8000==1.24.1